requirements.txt

Django==3.2.16
liffpy
line-bot-sdk

